added test for .fasta and .fasta.gz databases
