+++ Test/SearchEngineTests.cs
@@ -908,6 +908,7 @@ namespace Test
                 Assert.IsTrue(kvp.Value);
             }
         }
+
         #endregion Public Methods
     }
 }
